enhance(lang-service): performance improve for generate virtual code
