+++ packages/compiler/src/template/compose.ts
@@ -1,5 +1,5 @@
 import type { SourceLocation as BabelSourceLocation, ExportNamedDeclaration, ImportDeclaration, Node } from '@babel/types'
-import type { AttributeNode, BindingTypes, CodegenResult, CompilerOptions, NodeTransform, SourceLocation as VueSourceLocation } from '@vue/compiler-dom'
+import type { AttributeNode, CodegenResult, CompilerOptions, NodeTransform, SourceLocation as VueSourceLocation } from '@vue/compiler-dom'
 import type { VineCompFnCtx, VineCompilerHooks, VineCompilerOptions, VineFileCtx } from '../types'
 import { isExportNamedDeclaration, isFunctionDeclaration, isIdentifier, isImportDeclaration, isImportDefaultSpecifier, isImportSpecifier } from '@babel/types'
 import { compile, ElementTypes, NodeTypes, parse } from '@vue/compiler-dom'
@@ -17,19 +17,6 @@ const SHOULD_ADD_SUFFIX_REGEXP = /(?<=<[^>/]+)$/
 
 function toPascalCase(str: string) {
   return str.replace(/(?:^|-)(\w)/g, (_, c) => c.toUpperCase())
-}
-export function postProcessForRenderCodegen(codegen: string): string {
-  return codegen
-    // https://github.com/vue-vine/vue-vine/issues/171
-    // Replace all `= _resolveComponent('...')`, '...' is the component name,
-    // to `= (typeof <toPascalCase('...')> === 'undefined' ? _resolveComponent('...') : <toPascalCase('...')>)`
-    .replace(
-      /=\s*_resolveComponent\(['"](.+?)['"]\)/g,
-      (match, componentName) => {
-        const pascalComponentName = toPascalCase(componentName)
-        return `= (typeof ${pascalComponentName} === 'undefined' ? _resolveComponent('${componentName}') : ${pascalComponentName});`
-      },
-    )
 }
 
 function getTransformNegativeBoolPlugin(
@@ -58,67 +45,6 @@ function getTransformBareAttrAsBoolPlugin(
   )
 }
 
-export function compileVineTemplate(
-  vineCompFnCtx: VineCompFnCtx,
-  compilerHooks: VineCompilerHooks,
-  params: Partial<CompilerOptions>,
-  { ssr, getParsedAst = false }: {
-    ssr: boolean
-    getParsedAst?: boolean
-  },
-): (
-  CodegenResult
-  & { templateParsedAst?: ReturnType<typeof parse> }
-) | null {
-  const _compile = ssr ? ssrCompile : compile
-  const {
-    __enableTransformAssetsURL = true,
-    __enableTransformBareAttrAsBool = { transformNegativeBool: true },
-    __shouldAddTemplateSuffix,
-  } = compilerHooks.getCompilerCtx()
-    ?.options
-    ?.vueCompilerOptions ?? {}
-
-  try {
-    // vue/language-tools / #4583:
-    if (
-      __shouldAddTemplateSuffix
-      && SHOULD_ADD_SUFFIX_REGEXP.test(vineCompFnCtx.templateSource)
-    ) {
-      vineCompFnCtx.templateSource += '>'
-    }
-
-    return {
-      ..._compile(vineCompFnCtx.templateSource, {
-        mode: 'module',
-        hoistStatic: true,
-        cacheHandlers: true,
-        prefixIdentifiers: true,
-        inline: true,
-        nodeTransforms: [
-          ...getTransformBareAttrAsBoolPlugin(__enableTransformAssetsURL),
-          ...getTransformNegativeBoolPlugin(__enableTransformBareAttrAsBool),
-        ],
-        ...params,
-      }),
-      templateParsedAst: (
-        getParsedAst
-          ? parse(vineCompFnCtx.templateSource, {
-              parseMode: 'base',
-              prefixIdentifiers: true,
-              expressionPlugins: [
-                'typescript',
-              ],
-            })
-          : (void 0)
-      ),
-    }
-  }
-  catch {
-    return null
-  }
-}
-
 interface DefaultImportSpecifierMeta { type: 'defaultSpecifier', localName: string }
 interface NamespaceImportSpecifierMeta { type: 'namespaceSpecifier', localName: string }
 export interface NamedImportSpecifierMeta { type: 'namedSpecifier', specs: Map<string, string> }
@@ -134,7 +60,6 @@ export interface TemplateCompileComposer {
     vineFileCtx: VineFileCtx
     vineCompFnCtx: VineCompFnCtx
     mergedImportsMap: MergedImportsMap
-    bindingMetadata: Record<string, BindingTypes>
   }) => string
   templateCompileResults: WeakMap<VineCompFnCtx, string>
   generatedPreambleStmts: WeakMap<VineCompFnCtx, string[]>
@@ -309,6 +234,110 @@ function setVineTemplateAst(
   })
 }
 
+export function createTemplateCompileOptions(
+  vineFileCtx: VineFileCtx,
+  vineCompFnCtx: VineCompFnCtx,
+  compilerHooks: VineCompilerHooks,
+  inline: boolean = true,
+): Partial<CompilerOptions> {
+  return {
+    inline,
+    scopeId: `data-v-${vineCompFnCtx.scopeId}`,
+    bindingMetadata: vineCompFnCtx.bindings,
+    ...compilerHooks.getCompilerCtx()?.options?.vueCompilerOptions ?? {},
+    onError: (e) => {
+      compilerHooks.onError(
+        vineErr(
+          { vineFileCtx, vineCompFnCtx },
+          {
+            msg: `[Vine template compile error] ${e.message}`,
+            location: e.loc && computeTemplateErrLocation(vineFileCtx, vineCompFnCtx, e.loc),
+            rawVueTemplateLocation: e.loc,
+          },
+        ),
+      )
+    },
+    onWarn: (e) => {
+      compilerHooks.onWarn(
+        vineWarn(
+          { vineFileCtx, vineCompFnCtx },
+          {
+            msg: `[Vine template compile warning] ${e.message}`,
+            location: e.loc && computeTemplateErrLocation(vineFileCtx, vineCompFnCtx, e.loc),
+            rawVueTemplateLocation: e.loc,
+          },
+        ),
+      )
+    },
+  }
+}
+
+export function compileVineTemplate(
+  vineCompFnCtx: VineCompFnCtx,
+  compilerHooks: VineCompilerHooks,
+  params: Partial<CompilerOptions>,
+  { ssr, getParsedAst = false }: {
+    ssr: boolean
+    getParsedAst?: boolean
+  },
+): (
+  CodegenResult
+  & { templateParsedAst?: ReturnType<typeof parse> }
+) | null {
+  const _compile = ssr ? ssrCompile : compile
+  const {
+    __enableTransformAssetsURL = true,
+    __enableTransformBareAttrAsBool = { transformNegativeBool: true },
+    __shouldAddTemplateSuffix,
+  } = compilerHooks.getCompilerCtx()
+    ?.options
+    ?.vueCompilerOptions ?? {}
+
+  try {
+    // vue/language-tools / #4583:
+    if (
+      __shouldAddTemplateSuffix
+      && SHOULD_ADD_SUFFIX_REGEXP.test(vineCompFnCtx.templateSource)
+    ) {
+      vineCompFnCtx.templateSource += '>'
+    }
+
+    const compileResult = {
+      ..._compile(vineCompFnCtx.templateSource, {
+        mode: 'module',
+        hoistStatic: true,
+        cacheHandlers: true,
+        prefixIdentifiers: true,
+        inline: true,
+        nodeTransforms: [
+          ...getTransformBareAttrAsBoolPlugin(__enableTransformAssetsURL),
+          ...getTransformNegativeBoolPlugin(__enableTransformBareAttrAsBool),
+        ],
+        ...params,
+      }),
+      templateParsedAst: (
+        getParsedAst
+          ? parse(vineCompFnCtx.templateSource, {
+              parseMode: 'base',
+              prefixIdentifiers: true,
+              expressionPlugins: [
+                'typescript',
+              ],
+            })
+          : (void 0)
+      ),
+    }
+
+    // Store the template AST
+    setVineTemplateAst(vineCompFnCtx, compileResult)
+
+    return compileResult
+  }
+  catch {
+    return null
+  }
+}
+
 export function createSeparatedTemplateComposer(
   compilerHooks: VineCompilerHooks,
   ssr: boolean,
@@ -323,43 +352,17 @@ export function createSeparatedTemplateComposer(
       vineFileCtx,
       vineCompFnCtx,
       mergedImportsMap,
-      bindingMetadata,
     }) => {
       let hasTemplateCompileErr = false
       const compileResult = compileVineTemplate(
         vineCompFnCtx,
         compilerHooks,
-        {
-          scopeId: `data-v-${vineCompFnCtx.scopeId}`,
-          inline: false,
-          bindingMetadata,
-          ...compilerHooks.getCompilerCtx()?.options?.vueCompilerOptions ?? {},
-          onError: (e) => {
-            hasTemplateCompileErr = true
-            compilerHooks.onError(
-              vineErr(
-                { vineFileCtx, vineCompFnCtx },
-                {
-                  msg: `[Vine template compile error] ${e.message}`,
-                  location: e.loc && computeTemplateErrLocation(vineFileCtx, vineCompFnCtx, e.loc),
-                  rawVueTemplateLocation: e.loc,
-                },
-              ),
-            )
-          },
-          onWarn: (e) => {
-            compilerHooks.onWarn(
-              vineWarn(
-                { vineFileCtx, vineCompFnCtx },
-                {
-                  msg: `[Vine template compile warning] ${e.message}`,
-                  location: e.loc && computeTemplateErrLocation(vineFileCtx, vineCompFnCtx, e.loc),
-                  rawVueTemplateLocation: e.loc,
-                },
-              ),
-            )
-          },
-        },
+        createTemplateCompileOptions(
+          vineFileCtx,
+          vineCompFnCtx,
+          compilerHooks,
+          false,
+        ),
         {
           ssr,
           // Separate mode needs template's original AST,
@@ -370,9 +373,6 @@ export function createSeparatedTemplateComposer(
       if (!compileResult) {
         return ''
       }
-
-      // Store the template AST
-      setVineTemplateAst(vineCompFnCtx, compileResult)
 
       if (hasTemplateCompileErr) {
         return ''
@@ -432,7 +432,7 @@ export function createSeparatedTemplateComposer(
       // is combining all the bindings from user imports and all declarations.
       // non-inline mode, or has manual render in normal <script>
       // return bindings from script and script setup
-      const allBindings: Record<string, any> = { ...bindingMetadata }
+      const allBindings: Record<string, any> = { ...vineCompFnCtx.bindings }
       for (const key in vineFileCtx.userImports) {
         const isType = vineFileCtx.userImports[key].isType
         const isUsedInTemplate = vineFileCtx.userImports[key].isUsedInTemplate?.(vineCompFnCtx)
@@ -450,7 +450,7 @@ export function createSeparatedTemplateComposer(
       for (const key in allBindings) {
         if (
           allBindings[key] === false
-          || (bindingMetadata[key] === VineBindingTypes.PROPS)
+          || (vineCompFnCtx.bindings[key] === VineBindingTypes.PROPS)
         ) {
           // skip unused bindings
         }
@@ -463,7 +463,7 @@ export function createSeparatedTemplateComposer(
           // skip vue imports since we know they will never change
           setupFnReturns += `get ${key}() { return ${key} }, `
         }
-        else if (bindingMetadata[key] === VineBindingTypes.SETUP_LET) {
+        else if (vineCompFnCtx.bindings[key] === VineBindingTypes.SETUP_LET) {
           // local let binding, also add setter
           const setArg = key === 'v' ? '_v' : 'v'
           setupFnReturns += (
@@ -495,45 +495,17 @@ export function createInlineTemplateComposer(
       vineFileCtx,
       vineCompFnCtx,
       mergedImportsMap,
-      bindingMetadata,
     }) => {
       let hasTemplateCompileErr = false
       const compileResult = compileVineTemplate(
         vineCompFnCtx,
         compilerHooks,
-        {
-          scopeId: `data-v-${vineCompFnCtx.scopeId}`,
-          bindingMetadata,
-          ...compilerHooks.getCompilerCtx()?.options?.vueCompilerOptions ?? {},
-          onError: (e) => {
-            if (hasTemplateCompileErr) {
-              return
-            }
-            hasTemplateCompileErr = true
-            compilerHooks.onError(
-              vineErr(
-                { vineFileCtx, vineCompFnCtx },
-                {
-                  msg: `[Vine template compile error] ${e.message}`,
-                  location: e.loc && computeTemplateErrLocation(vineFileCtx, vineCompFnCtx, e.loc),
-                  rawVueTemplateLocation: e.loc,
-                },
-              ),
-            )
-          },
-          onWarn: (e) => {
-            compilerHooks.onWarn(
-              vineWarn(
-                { vineFileCtx, vineCompFnCtx },
-                {
-                  msg: `[Vine template compile warning] ${e.message}`,
-                  location: e.loc && computeTemplateErrLocation(vineFileCtx, vineCompFnCtx, e.loc),
-                  rawVueTemplateLocation: e.loc,
-                },
-              ),
-            )
-          },
-        },
+        createTemplateCompileOptions(
+          vineFileCtx,
+          vineCompFnCtx,
+          compilerHooks,
+          true,
+        ),
         { ssr },
       )
       if (!compileResult) {
@@ -541,9 +513,6 @@ export function createInlineTemplateComposer(
       }
 
       const { preamble, code } = compileResult
-
-      // Store the template AST
-      setVineTemplateAst(vineCompFnCtx, compileResult)
 
       if (hasTemplateCompileErr) {
         return ''
@@ -579,4 +548,18 @@ export function createInlineTemplateComposer(
       return finalCode
     },
   }
+}
+
+export function postProcessForRenderCodegen(codegen: string): string {
+  return codegen
+    // https://github.com/vue-vine/vue-vine/issues/171
+    // Replace all `= _resolveComponent('...')`, '...' is the component name,
+    // to `= (typeof <toPascalCase('...')> === 'undefined' ? _resolveComponent('...') : <toPascalCase('...')>)`
+    .replace(
+      /=\s*_resolveComponent\(['"](.+?)['"]\)/g,
+      (match, componentName) => {
+        const pascalComponentName = toPascalCase(componentName)
+        return `= (typeof ${pascalComponentName} === 'undefined' ? _resolveComponent('${componentName}') : ${pascalComponentName});`
+      },
+    )
 }
